consolidate errors to errors.py
