+++ hwilib/cli.py
@@ -2,8 +2,13 @@
 
 from .commands import backup_device, displayaddress, enumerate, find_device, \
     get_client, getmasterxpub, getxpub, getkeypool, prompt_pin, restore_device, send_pin, setup_device, \
-    signmessage, signtx, wipe_device, NO_DEVICE_PATH, DEVICE_CONN_ERROR, NO_PASSWORD, \
+    signmessage, signtx, wipe_device
+from .errors import (
+    NO_DEVICE_PATH,
+    DEVICE_CONN_ERROR,
+    NO_PASSWORD,
     UNKNWON_DEVICE_TYPE
+)
 
 import argparse
 import getpass
